GUI + Animation Functionality
The user can now generate plots extremely quickly through a GUI. Saving plots as .png's is now an option, as well as a slightly experimental "animation mode". By checking the "Animate mode" button and clicking update, a series of frames is saved to an "Animation Frames" folder. By using a program such as Ffmpeg, a video file can be made. Currently the frame rate and the exact parameters of the movements within the animation are hardcoded, but the resolution is user-defined.

I also included a functionality to print out an estimated render time in an hour-minute-second format so the user will know how long their animation will take to make. The estimations assume the time between each frames' rendering is uniform. So far, the estimations have been extremely accurate.

An "Animation Frames" and "Saved Plots" folder may be needed for this code to work.

--- Known Bugs ---

- An oddly thick line sometimes appears. This is likely a ray that is caught between two parallel reflective surfaces, causing an overlapping reflection to occur. This could be fixed with some way to detect when such an event occurs.