Render Estimator Upgrade
The code now uses a more robust way of estimating render times by curve-fitting a small number of frame render times (20-50) to a quadratic. The accuracy can be changed by increasing the number of sampled points. A current bug exists where negative times can result due to the quadratic being concave-down. Obviously the render times can decrease as the frame number increases, but it should not decrease to or past 0. A few more animation options are included and the GUI's code was made more clear by using for loops instead of giant lists of objects.